+++ FirebaseAI/Tests/Unit/TestUtilities/GenerativeModelTestUtil.swift
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import FirebaseAppCheckInterop
+import FirebaseAuthInterop
+import FirebaseCore
 import Foundation
 import XCTest
 
@@ -74,6 +77,44 @@ enum GenerativeModelTestUtil {
         return (response, fileURL.lines)
       }
     #endif // os(watchOS)
+  }
+
+  static func nonHTTPRequestHandler() throws -> ((URLRequest) -> (
+    URLResponse,
+    AsyncLineSequence<URL.AsyncBytes>?
+  )) {
+    // Skip tests using MockURLProtocol on watchOS; unsupported in watchOS 2 and later, see
+    // https://developer.apple.com/documentation/foundation/urlprotocol for details.
+#if os(watchOS)
+    throw XCTSkip("Custom URL protocols are unsupported in watchOS 2 and later.")
+#endif // os(watchOS)
+    return { request in
+      // This is *not* an HTTPURLResponse
+      let response = URLResponse(
+        url: request.url!,
+        mimeType: nil,
+        expectedContentLength: 0,
+        textEncodingName: nil
+      )
+      return (response, nil)
+    }
+  }
+
+  static func testFirebaseInfo(appCheck: AppCheckInterop? = nil,
+                                auth: AuthInterop? = nil,
+                                privateAppID: Bool = false) -> FirebaseInfo {
+    let app = FirebaseApp(instanceWithName: "testApp",
+                          options: FirebaseOptions(googleAppID: "ignore",
+                                                   gcmSenderID: "ignore"))
+    app.isDataCollectionDefaultEnabled = !privateAppID
+    return FirebaseInfo(
+      appCheck: appCheck,
+      auth: auth,
+      projectID: "my-project-id",
+      apiKey: "API_KEY",
+      firebaseAppID: "My app ID",
+      firebaseApp: app
+    )
   }
 }
